Update package to use the DarkSky API
The forecast.io API has been shelved, and is now merging into the
DarkSky API.

The DarkSky API is semantically equivilant to forecast.io's. The two
were essentially sister websites, and the two have now consolidated.

This involves a simple renaming of the package, renaming any links to
forecast.io to darksky.net, and changing the API endpoint.